+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 ## Changelog
+
+- [anti_build] feature: new `attack_entity` anti build type support.
+  - **NOTE: The `atatck_entity` come with `minecraft:villager` as its default config. You may need to configure it, to satisfy your needs.**